backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pandas
python-dotenv
google-genai
requests
nba_api
cachetools
python-multipart
